data-generator/requirements.txt
kafka-python==2.0.2
faker==22.0.0
python-dateutil==2.8.2
